Only call save_mysql_dbs if using mysql
This adds some conditional logic around when mysql
is used based on the configured database type in
the devstack rc file. TODOs are left in place for
supporting postgresql which can be fleshed out when
we have a grenade-postgresql CI job to verify the
changes.

+++ clean.sh
@@ -72,10 +72,14 @@ ps auxw | grep ^rabbitmq | awk '{print $2}' | sudo xargs kill
 sudo service rabbitmq-server stop
 sudo apt-get purge -y rabbitmq-server .*erlang
 
-# Get ruthless with mysql
-service mysqld stop
-sudo apt-get  purge -y .*mysql-server
-sudo rm -rf /var/lib/mysql
+if [[ "${DATABASE_TYPE}" == "mysql" ]]; then
+    # Get ruthless with mysql
+    service mysqld stop
+    sudo apt-get  purge -y .*mysql-server
+    sudo rm -rf /var/lib/mysql
+else
+    echo "TODO: cleanup ${DATABASE_TYPE}"
+fi
 
 # kill off swift, which doesn't live in screen, so doesn't die in screen
 ps auxw | grep swift | awk '{print $2}' | xargs kill
